Add DiffSpec endpoint model
